feat: add strictDependencies option when deploying datapacks
With strict dependencies enabled the deployment order is changed to ensure datapacks and all direct dependencies are deployed before deploying a datapack that depends on it. This results in a less optimal deployment chunking but does help to ensure datapack records aren't referenced by other datapacks until they are fully deployed.