feat: introduce new uniform query API for data and tooling objects; update query service to use new query API. The query API is supported through `query2` method on the salesforceConnection and has better support for querying related objects and nested resources if not returned in the main result set. Query2 response implements a new async-iterator and is backward compatible with jsforce's event emitting query class ahwile also supporting promise like behavior for awaiting the whole result set.
Other functions fo the query service such as record wrapping have moved to a new factory class which makes it easier to use and test the functionality of the wrapper. Thw wrapper record allows accessing fields using normalized field names making it easier to directly work with records returned in queries. The factory supports 2 strategies; either wrapping the record in a proxy or pre-defining all properties in their normalized variant on the object.